+++ docs/source/changes.md
@@ -293,6 +293,8 @@ authors:
 - Environment variables are no longer used to control execution and variables,
   use command-line switches instead
   ({{ gh(663) }} by {{ authors.larsoner}} )
+- Misspellings are now checked in config
+- ({{ gh(665) }} by {{ authors.larsoner }} )
 
 ### Code health